Updated to include pubsub and mqtt pubsub functions (#86)
* Updated to include pubsub and mqtt pubsub functions

I still have a question about the sample code in the MQTT publish section. However, I think the rest of the content is good to go based on the comments I saw in the review.

* Modified MQTT publish sample

Updated the first section of the sample application to match the sample in MQTT subscribe.